multi-conn: New filter
Implement a TODO item of emulating multi-connection flush consistency
via multiple plugin flush calls to allow a client to assume that a
flush on a single connection is good enough.  This also gives us some
fine-tuning over whether to advertise the bit, including some setups
that are unsafe but may be useful in timing tests.

Testing is interesting: I used the sh plugin to implement a server
that intentionally keeps a per-connection cache.

Note that this filter assumes that multiple connections will still
share the same data (other than caching effects); effects are not
guaranteed when trying to mix it with more exotic plugins like info or
tmpdisk that violate that premise.