file: Better support for read-only files
Previously, attempts to use nbdkit without -r in order to visit a
read-only image failed to open the image (in fact, a single read-only
file in an otherwise usable directory makes "nbdkit file dir=. --run
'nbdinfo --list'" fail to print anything in libnbd 1.6).  But a saner
approach is to try a fallback to a readonly fd if a read-write fd
fails.

The windows code passed under Wine (thanks: Rich Jones), where we also
have to disable flush support on readonly files.