filters: Slightly reduce promise of nxdata lifetime
We promise (and test-layers verifies) that the pointer value of nxdata
passed to everything from .prepare to .finalize will be the same.  But
the wording was ambiguous on whether that promise extended to the
nxdata of .open as well.  In practice, this had been possible
previously, but will get harder in upcoming patches that want to
switch over to a struct context* that is not determined until after
the next layer has been opened.  So update the wording and test to
match our intent.