ext2: Override plugin .can_trim/.can_zero/.can_flush
Since we didn't override .can_trim or .can_zero, if the plugin
supports that and we end up advertising it to the client, then the
client trying to take advantage of them would end up trimming or
zeroing at the wrong offsets in the plugin, probably corrupting the
image.  (Thankfully not an issue for nbdkit -r when you're just
reading the file system, since nbdkit doesn't advertise trim or zero
on read-only connections).

If we don't override .can_flush, we are at the mercy of the plugin for
whether the client will know that it can stabilize things into the
ext2 filesystem (regardless of whether we can then further stabilize
that filesystem back to the underlying plugin storage).

Finally, the docs for nbdkit-filter say we should not call
next_ops->cache unless can_cache returned NBDKIT_CACHE_NATIVE, but we
were checking for equality to 1 (NBDKIT_CACHE_EMULATE).

Fixes: 86b8466d260 (filters: Add ext2 filter)
(cherry picked from commit 361c192f0e96ac365ae7187f5089630dc558a7df)