ext2: Simplify use of next_ops
We no longer need a 'struct nbdkit_next' now that a single pointer is
sufficient for calling into the backend.

There's still more to do with the filter to allow multiple client
plugins to share a single context into the plugin, but that requires a
bit more work (for starters, we need a way to clean up that single
context; .unload is too late, but doing it in .close requires tracking
how many open client connections there are).