filters: Temporarily make next_ops->reopen a public function
The reopen operation (key to the retry filter) is enough of an oddball
that it does not belong as a function pointer in next_ops; in
particular, while most functions in next_ops operate on an existing
context, reopen can operate even after a previous failure has left no
context, but passing NULL for nxdata to represent no context does not
give us the information we need to associate the reopened context back
to the right connection.  Wrap backend_reopen with a new exported
function nbdkit_backend_reopen, and drop the next_ops->reopen function
pointer.  Note that the function is only declared for the retry filter
at this time, and that while at it, I made the retry filter use a bit
more type safety.  Later patches will perform more mechanical
conversion of other filters to be type-safe.

The new function depends on guaranteeing that the 'nxdata' passed to
.open is stable for usage as the 'backend' parameter to pass to
nbdkit_backend_reopen, so I renamed that parameter in the docs.  Our
testsuite coverage of the retry filter ensures we don't break this.
And it gets us one step closer to allowing plugins to open a context
into the backend without needing it to be tied to the current
connection from the client.  In this patch, 'backend' of .open and
'nxdata' of .pread happen to be the same pointer and reopen doesn't
affect it, but in the upcoming patches, the two will be distinct
pointers, where nxdata can sometimes be NULL on entry to .pread and
set back to non-NULL by reopen.

Note that the new function is intended to be temporary while I
continue to make other refactorings.  Ultimately, I want to get to the
point where any filter can open up a new context without regards to
the current connection, and then optionally associate any context into
the current connection (that is, the next_open callback passed to
.open will become shorthand for the paired action of
nbdkit_backend_open/nbdkit_backend_associate); once that is in place,
we don't need nbdkit_backend_reopen because the retry filter can just
directly open new connections as needed.