server: Update signature of backend_open
Instead of returning an int on success, return a pointer to everything
needed to call into the next layer via next_ops.  The public interface
in nbdkit-filter.h still returns int, so we need a next_open() wrapper
added back in filters.h.  Similarly, an upcoming patch wants to update
backend_* functions to directly take a struct context* instead of the
current approach of a struct backend* coupled with querying TLS for
the current connection.  This will be easier by making struct context
track the struct backend that it was opened under.  Struct connection
still maintains a list of contexts, but now instead of directly
populating direct members of an array during backend_open, we now
populate pointers in the callers of backend_open; this will let us
progress towards filters opening a context independently from the
current client connection.