server: Store nbdkit_next_ops in struct context
Time to take advantage of C typing rules.  First, make 'struct
context' a derived class with 'struct nbdkit_next_ops' as its first
member, so that a pointer to one is now the same pointer to the other
at the ABI level.  Next, note that all entries in filters.h's next_ops
are backend_* functions, so we can move that static list into
backend.h instead, and populate next_ops as part of backend_open.
Since filters still see 'nbdkit_context*' typedef'd to 'void*', they
are none the wiser, even though filters.c is now passing the same
pointer to two different parameters (since &c_next->next == c_next).
Yes, that means that next_ops->FOO(nxdata) can now be equally spelled
as next_ops->FOO(next_ops).

The oddball (once again) is the retry filter; since
nbdkit_backend_reopen only affects one of the two pointers, the filter
is now updated to ignore one parameter in favor of the other.

A later patch will then do the mechanical change to all the remaining
filters to be fully type-safe at the same time we reduce the parameter
duplication to just pass a single copy of context.  The upshot of this
refactoring is that it also becomes easier to introduce a new function
for opening a context from a filter, which can return a single pointer
containing everything the filter needs to call into the plugin.