Update github URLs to point to gitlab.
A few weeks ago we moved the hosting to gitlab.  The URL structure is
sufficiently similar that it was possible to use a Perl script to
mechanically substitute URLs pointing to github so they point to the
same location on the new gitlab instance.

For the announcement of the move:

https://listman.redhat.com/archives/libguestfs/2021-February/msg00021.html