multi-conn: Add knob to limit flush replication by export name
Many plugins serve the same content regardless of export name; on such
plugins, a flush operation on one connection must affect all others.
But some plugins serve distinct content on different export names; on
those, replicating a flush operation to another connection visiting a
different export is wasted work (but thankfully, not semantically
incorrect, even if the two exports have different sizes, because
NBD_CMD_FLUSH does not take a size and offset, and flush does not
corrupt data).  We want to add a knob to optimize the latter, even
though the knob must default to off because of the former.

Done by adding another layer of indirection: all connections are
combined into a group (a singleton when the knob is off, or by name
when the knob is on), with fast dirty tracking moved from file scope
into group scope.