multi-conn: Multiple updates to the documentation.
Improve the title to make it non-generic.

Start by introducing the feature and motivation for the filter.

The flag is NBD_FLAG_CAN_MULTI_CONN (not NBD_FLAG_MULTI_CONN), but I
only mention the full name once, abbreviating it to "multi-conn"
elsewhere.

Multiple other smaller changes to improve readability.