tests: Work around nbdinfo 1.6.2 bug
We had a release of libnbd that failed to exit with non-zero status on
a connection error.  Although the libnbd regression has been fixed in
the meantime, it's easy enough to tweak the nbdkit testsuite to work
around the broken releases rather than waiting until the fixed version
has hit all the distros.