curl: Make sure all constants passed to curl_easy_setopt are long.
The function is declared in <curl/easy.h> as varargs:

  CURL_EXTERN CURLcode curl_easy_setopt(CURL *curl, CURLoption option, ...);

so we must be very careful to pass the correct type since C will not
promote it automatically.  In several places we were passing an int
instead of a long.

(cherry picked from commit cd3a318feec58484b1384447ba527733147970a5)