tests/test-captive.sh: Use a different signal to avoid core dump
For a long time this test would leave a core dump file around.  I
thought this only happened under valgrind (leaving a "vgcore" file).
However I realised recently that this was because systemd-coredump
had been “hiding” the non-valgrind core dumps from me, and in fact
they were still happening even for regular “make check”.

The fundamental issue with the test was the use of SIGABRT which is
designed to core dump if not caught, so essentially the kernel was
doing exactly what we asked.

To fix this I have changed the signal to SIGUSR1.  This signal is also
unhandled by nbdkit so we should be able to observe its effect on the
exit code of the captive nbdkit, and it has the advantage that the
default action of SIGUSR1 is not to core dump.

This commit also enables the test again under valgrind, reverting the
change made in commit fe2bf4037d8f1ea238aa49ab0a048a8e55985057.

Reverts: commit fe2bf4037d8f1ea238aa49ab0a048a8e55985057
Fixes: commit 38165d0f8f5a4eb898d756a561811c2fe042fe4d