tar: Add comment about why we use LANG=C
virt-v2v uses some homebrew code to parse tar output (it should be
using nbdkit-tar-filter).  That code failed because the output from
tar depends on language, so we had to fix it in virt-v2v by setting
LANG=C.

While investigating the virt-v2v bug, I checked that nbdkit-tar-filter
already sets LANG=C, which it does apparently for hygiene reasons and
not because we hit the same bug before.

Adding a comment to the code here to explain why it is necessary.