filters: Tighter rules on .get_ready/.after_fork
Allowing a filter to bypass the plugin's .get_ready or .after_fork
makes no sense, as there are no other arguments to modify.
Furthermore, now that filters have access to open a plugin context
independently of a client connection, .after_fork would be an ideal
place to do so with its existing nbdkit_backend argument (which we
already documented was stable, although this commit actually adds
tests for that).  Still, we will want to add a counterpart API (in a
later patch) to tear that context down cleanly, where .unload is too
late.  But for that to work, we need to ensure that the plugin
completes initialization before the filter is allowed to try to open a
context into the plugin, which means swapping the visitation order to
be inner-to-outer (similar to .prepare).