threadlocal: Add way to store current context
Although we've recently added nbdkit_next_context_open to make it
easier to access a plugin from a filter, to date we have still been
limited to opening a context while associated with a client
connection, because backend_open uses GET_CONN which asserts there is
an active connection.  In order to lift that restriction, we will need
to be able to allow a context that has no associated connection, and
move the data that feeds nbdkit_export_name() and friends out of the
connection and into the context.  This patch starts the process, by
updating the thread local storage to track the current context, as
well as all calls into backend.c to keep this information up-to-date.

As this is internal to the server, I did not think it proper to add
this into common/utils/cleanup.h, but rather just placed the new
macros in internal.h.