api: Add .cleanup callback
The existing .unload callback is currently called inner-to-outer,
which is not suitable for use in late cleanups where a filter wants to
make one final call into the underlying plugin.  We were careful to
state that unload does not have a specific ordering, but switching
that ordering may not be helpful.  Late cleanup may also want one
final chance to open a context into the plugin, which implies one
final access to the backend pointer from filters (in contrast to
.unload taking no parameters).

So, I found it easier to introduce a new .cleanup callback,
specifically stated to be the counterpart of .after_fork.  This leaves
.load/.unload as a pair for handling things at the dlopen barrier
(affecting only the local filter or plugin, but no interaction with
other modules), and .after_fork/.cleanup for handling things related
to the first and last points at which it is safe to interact with the
underlying plugin.  The new callback has no return value, as by the
time we are cleaning up, any failures are not going to change the fact
we are shutting down anyway.