filters: Allow nbdkit_next_context_open outside client connection
Until now, attempts to create a new context outside a client
connection (such as during .after_fork) triggered an assertion
failure: backend_open insisted on having a valid connection.  Even if
we merely drop that assertion, there are other issues to worry about:
nbdkit_is_tls() depends on the current connection to know if the
client negotiated TLS; if the backend interns a string with
nbdkit_string_intern() while there is a current connection, then we
risk use-after-free since that string's lifetime ends when the client
disconnects even if the context stays around; and so on.  Since we
never promised filter ABI stability, we can address this by adding a
parameter to state the intent of whether the filter intends to share a
single backend context among multiple client connections.  Existing
uses all pass false (the context lifetime is tied to the client
connection, so life continues as before), but we now open the door to
a filter that multiplexes connections, where shared connections now
have a global interned string lifetime, a choice of TLS determined by
the command line, and better documentation.

The handling of nbdkit_is_tls is the most interesting aspect: if a
user were to mix a multiplexing filter in front of the tls-fallback
filter, we want to never expose data from the real plugin unless
--tls=require was on the command line, because otherwise we'd have a
lot more code to audit and touch to make sure the multiplexing filter
doesn't accidentally leak data to an insecure client.  But if the
tls-fallback filter is in front of the multiplexing filter, things
should just work: the multiplexing filter is not reached except by TLS
clients.