tests: Only set MALLOC_CHECK_ and MALLOC_PERTURB_ in tests.
Previously these variables were set in the wrapper.  However the
wrapper has a dual purpose, as both a place to run nbdkit without
installation, and a way to run the tests.  When benchmarking, we don't
want malloc-check being used because of the overhead, but we still
want to run everything from the local directory for fast iteration.

Solution is to move these variables to TESTS_ENVIRONMENT so they only
run from tests.

See similar commit in libnbd:
https://gitlab.com/nbdkit/libnbd/-/commit/88e72dcb1631b315957f5f98e3cdfcdd1fd0fe29