Set visibility to default for exported symbols
In plugins we expect the plugin_init function to be exported as well
as *_debug_* variables.  In the server various nbdkit_* functions
should be exported.

If you compile nbdkit with -fvisibility=hidden these symbols are not
exported at all even though the linker script says they should be.
This option can be useful as it allows the compiler to generate
simpler and hence faster code, and it's required for Clang CFI.

Set the visibility to "default" for all symbols we expect to be
exported.