tests: Don't break if libssh is installed but not ssh-keygen
If libssh is installed then we can build the ssh plugin.  However we
cannot test the plugin without both ssh-keygen and sshd.  We already
"requires sshd" so that was fine, but tests/Makefile.am used
ssh-keygen directly to build the test host keys etc and would break if
it was not installed.  Use a configure test to avoid trying to
building the test data in this case.

https://aur.archlinux.org/packages/nbdkit/#comment-811389

Reported-by: Thomas Weißschuh