docs: Remove nbd-client '-b 512' flag from documentation
In nbd-client <= 3.19 the default sector size was 1024 bytes which
caused peculiar problems with partitioning programs like fdisk and
parted.  This required you to use 'nbd-client -b 512' to get more
reasonable behaviour.

Because this has been fixed since the commit below in the nbd
repository (nearly 3 years ago), we no longer need to add this
workaround everywhere in nbdkit’s documentation.

  commit 128fd556286ff5d53c5f2b16c4ae5746b5268a64
  Author: Wouter Verhelst
  Date:   Wed Nov 28 21:47:38 2018 +0100

    Change blocksize to 512

    1024 bytes was originally chosen as fitting nicely inside an ethernet
    frame. That didn't really make much sense, but the issue was masked by
    the kernel actually ignoring block sizes due to a bug.

    Now that the kernel bug is fixed, the result is that partitions are
    parsed incorrectly, resulting in confusing behaviour.

    Fix by defaulting to 512. For the rare case where 1024 bytes still makes
    sense, the -b parameter still exists.