cow: Fix assert failure in cow_extents
$ nbdkit sparse-random 4G --filter=cow --run 'nbdinfo --map $uri'
nbdkit: cow.c:591: cow_extents: Assertion `count > 0' failed.

nbdinfo calls us with count = 0xfffffe00 [1] stored in a 32 bit
quantity.  This was rounded up to the next block boundary and so
overflows (count becomes 0, hence the assertion).

Use a 64 bit variable for count to allow rounding up.  This requires
further changes as a further 32 bit variable must not be allowed to
overflow.

This also adds a regression test.

[1] UINT32_MAX - 512 + 1 from:
https://gitlab.com/nbdkit/libnbd/-/blob/c55c5d9960809efd27cd044d007a33ea1636f4b0/info/map.c#L64