cache: Implement cache-on-read=/PATH
For virt-v2v we will need to be able to turn cache-on-read on while
performing inspection and modification of the guest, and off when
doing the bulk copy.  To do that allow the cache-on-read parameter to
refer to a path where the existence of the path toggles the feature.

(We could restart nbdkit between these phases, but this change avoids
doing that.)

(cherry picked from commit c8b575241b15b3bf0adaf15313e67e5ed4270b5a)