cache, cow: Use a 64K block size by default
Based on the results presented in the previous commit, use a 64K block
size by default in both the cache and cow filters.  For the cache
filter you could go back to a 4K block size if you wanted by using the
cache-min-block-size=4K parameter.  For cow it is compiled in so
cannot be adjusted.