server: Make debug messages atomic
Previously we used flockfile around the separate fprintf statements
used to emit debug messages.  However this does not make debug
messages atomic (only atomic with respect to other nbdkit threads).
It's easy to see this in strace:

1915655 write(2, "nbdkit: ", 8)         = 8
1915655 write(2, "debug: ", 7)          = 7
1915655 write(2, "nbdkit 1.27.3 (nbdkit-1.27.3-1.f"..., 36) = 36
1915655 write(2, "\n", 1)               = 1

We require open_memstream already, so use this to print messages to a
string buffer which we can emit in a single write.

This fixes various tests which grep the log file looking for sentinel
messages.  In particular a recent Fedora build failure in
test-nbdkit-backend-debug.sh was caused by this, but there are other
tests that fail occasionally for the same reason.

This change also tries harder to set errno correctly before expanding
the format string (for debug messages that contain %m).

Note that libnbd already emits atomic messages.