delay: Improve parsing and representation of delay times
We used sscanf to parse the delay settings for both NNms and NN
(seconds).  In the first case we need to use sscanf otherwise the code
is awkward.  In the second case using sscanf meant that bogus suffices
were not ignored, eg:

  $ nbdkit null --filter=delay delay-open=10SECS --run 'nbdinfo $uri'

Use nbdkit_parse_unsigned instead.  This now results in an error:

  nbdkit: error: delay-open: could not parse number: "10SECS": trailing garbage

Also, previously we stored delay times as an int.  This makes no sense
since they cannot be negative, so use an unsigned instead.

Reported-by: Ming Xie