server: nanosleep: Change error for early end of sleep
At the moment nbdkit_nanosleep gives an incorrect error message if it
aborts early.  Even in the case when the server is not actually
shutting down it will say:

  $ nbdkit --filter=delay null delay-close=3 --run 'nbdinfo --size $uri; nbdinfo --size $uri'
  0
  nbdkit: null[1]: error: aborting sleep to shut down
  0
  nbdkit: null[2]: error: aborting sleep to shut down

This commit changes the error so we only talk about shut down when the
server is actually shutting down, and use a different message in other
cases.