common/allocators/malloc.c: Remove bogus kernel hints
These kernel hints are wrong in several ways.

MADV_DONTFORK should not be used because when we use captive nbdkit
(the --run option) we do actually fork and run nbdkit as the child.
However the kernel does not have to provide the mallocd memory to this
child process so it disappears.

Even if the hints were not wrong, setting them using
madvise (ma->ba.ptr ...) would be wrong because the allocator buffer
can be extended at any time using realloc and could move in memory.
The hints would then apply to unrelated glibc allocations.  I believe
this is what caused the crash I observed.

For some reason the bug was only seen on s390x where it caused memory
corruption in glibc followed by a crash, but I don't believe this bug
is specific to s390x, it's just something about that architecture that
made it more likely to happen.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1992542
(cherry picked from commit 557a7a85c944dcd247feb0d670b0deca8da46576)