maint: Default to probing valgrind at runtime
When we added --enable-valgrind in commit ac4ad5545, we argued that a
runtime probe for valgrind might be invasive enough that it should be
opt-in for developers only.  But it turns out that valgrind.h's
implementation for RUNNING_ON_VALGRIND is intentionally minimally
invasive: it does not link in any external library or function calls,
but merely injects a few benign instructions (a sequence of rol on
x86_64) that the valgrind JIT recognizes as special, but which are a
no-op on bare-metal runs.  Furthermore, we are not checking valgrind
use in a hotspot.  So there is no real penalty to defaulting the
valgrind probe to on, which makes 'make check-valgrind' more likely to
work out-of-the-box.  Still, we keep --disable-valgrind for anyone
that does not want to use 'CFLAGS=-DNVALGRIND' as another alternative
for avoiding even those few assembly instructions in a production
binary.

Also, we misused AC_ARG_ENABLE, such that './configure
--disable-valgrind' actually enables valgrind.  We then copied that
same bug for --disable-libfuzzer (commit 9b0300ca).  An audit of the
other AC_ARG_WITH and AC_ARG_ENABLE did not find other instances of
this misuse.

In addition to tweaking the configure default, we need to rearrange
the logic of internal.h so that libfuzzer and sanitize address still
unconditionally disable dlclose, even when valgrind headers were
detected.

(cherry picked from commit be1fe9abb8e3a3448bd1deedf45f712736e7d8a5)