data: Add a new EXPR_FILL expression type to optimize <byte>*N
The manual mentions that <byte>*N can be used to repeat a byte.  This
was optimized already in the specific case where the left hand side
was a byte expression, but not in other cases (like a single character
string).  Also the optimization was done late.

Add a new EXPR_FILL expression type.  During AST optimization replace
some single byte expressions * N (including 1 byte strings) with a fill.