data: When possible, combine adjacent list elements
The existing optimizer could optimize lists of byte elements into
strings (eg. 1 2 3 => "\x01\x02\x03").  But if a list contained
anything else it would not be optimized, even if optimization of parts
of the list were possible.

Use a more generic optimization pass which tries to combine any
adjacent elements of the list.  This can handle other types, like
combining bytes and fills, or strings and strings, and also partially
optimizes if the list contains other elements that cannot be combined.