maint: Mention online gpg keyring for security use
As part of me taking on release duties, we expanded the keyring used
to validate Fedora builds; this keyring in turn happens to match the
set of interested developers for dealing with security bugs.  Since
various GPG keyservers have lately become harder to use due to
overpopulation of bogus keys, pointing to a non-keyserver solution is
thus a worthwhile idea.