cow: Make the block size configurable
Commit c1905b0a28 ("cache, cow: Use a 64K block size by default")
changed the nbdkit-cow-filter block size to 64K, but it was still a
fixed size.  In contrast the cache filter allows the block size to be
adjusted.

Allow the block size in this filter to be adjusted up or down with a
new cow-block-size=N parameter.

When using the VDDK plugin, adjusting this setting can make a
difference.  The following timings come from a modified virt-v2v which
sets cow-block-size and was used to convert from a VMware server to
-o null (this is also using cow-on-read=true):

  cow-block-size=64K:   18m18
  cow-block-size=256K:  14m13
  cow-block-size=1M:    14m19
  cow-block-size=4M:    37m33

As you can see it's not obvious how to choose a good block size, but
at least by allowing adjustment we can tune things.