data: Fix optimization of inlined lists
Commit 6cfbcf9002 ("data: Get rid of EXPR_EXPR") is generally correct
but made a mistake with inlining of lists.  Offsets are scoped and so
should not be inlined.  For example:

  1 (@2 2) (@3 3)

should expand to:

  1 0 0 2 0 0 0 3

but because the lists are inlined it currently expands to:

  1 0 2 3

Fix this optimization and add a test.

Note that this only worked in the old (EXPR_EXPR) code by accident
because I missed an optimization there.

Fixes: commit 6cfbcf900273fd444426bf446a5d96a874dc3e9d