data: Fix issues on 32 bit
Commit 9352ad00e3 ("data: New functions new_node/get_node with simpler
usage") added a new expr() function to let you construct expressions.
However it uses varargs so callers have to be strict about the types
of the parameters, and the compiler cannot verify this.

When creating EXPR_FILL the parameters are uint8_t, uint64_t, where
varargs always promotes uint8_t to int.  We called it in a few places
with incorrect types.  The error showed up when compiling on i686, but
actually we were very lucky this did not manifest on 64 bit platforms
too.  It only worked because of how x86-64 ABI passes some arguments
in registers so that int is passed in a 64 bit register.  On a 64 bit
platform that uses the stack it could have failed.

Use casts as appropriate.  Strongly typed qualified unions à la OCaml
would help here!

Fixes: commit 9352ad00e3e1fff679f4eabb2208bfa1e1443a29