data: Fix inlining of unsafe expressions
Commit 9bea4b396e ("data: Fix optimization of inlined lists") fixed
inlining of lists in lists.  However it missed the case where a single
element is inlined in the superior list.

An example which was optimized wrongly was:

  ( 0 -> \a ) \a

This was incorrectly optimized by inlining to:

  0 -> \a \a

However it should have failed (because the use of \a is out of scope
of its definition).

This commit corrects this and also adds a general test of some
expressions that should fail to parse or be optimized.

Fixes: commit 9bea4b396edd350f13a4c57e455c482a9853e415
Fixes: commit 6cfbcf900273fd444426bf446a5d96a874dc3e9d