cow: Fix for qemu 6.1 which requires backing format
The diffing example in the manual created a qcow2 file with a backing
file but did not specify the backing format.  However qemu 6.1 now
requires this and fails with:

  qemu-img: cow-diff.qcow2: Backing file specified without backing format

or:

  qemu-img: Could not change the backing file to 'cow-base.img': backing format must be specified

Fix the example by adding the -F option to the command line.

Also there was a test of this rebasing sequence which failed, so this
commit updates the test too.

(cherry picked from commit 618290ef33ce13b75c1a79fea1f1ffb327b5ba07)