common/include/checked-overflow.h: Simplify
We don't need separate macros for size_t and uint64_t since the compiler
builtins are generic. If we need to port to a complier that does not have
generic builtins, it is likely that we can do this in a generic way.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from libnbd commit 46fa6ecc7422e830f10d64b62c808abe6f8b49af)