tests: Remove redundant comment
Commit e3abac4e5519b0d51c9a34a41af97f63920f3f32 which originally added
the OCaml plugin back in 2014 added this comment.  At the time the
test was novel because it compiled a plugin (instead of testing either
an existing plugin or one of the example plugins which is designed to
testing against).  However lots of other tests do that now so the
comment is redundant.

Thanks: Laszlo Ersek
(cherry picked from commit 6b821fdc4ced4fc0af3fe2480d3ca43cb6476a6d)