ocaml: Replace plugin struct with register_plugin labelled parameters
Previously in OCaml plugins you had to create a "plugin struct".  This
looked a bit like - but was entirely unrelated to - the nbdkit_plugin
struct in C plugins.  We copied fields from this phony struct into the
real C struct at run time.

There were several problems with this, the main one being the awkward
way that required and optional fields were not resolved until run
time.  This could result in run time only failures (so the plugin was
not fully compile time safe).  Another problem was confusion over ""
vs value not present, because "" was used as a sentinel to mean not
present.

We can instead use labelled parameters, which have a natural and
compile time safe way to specify optional parameters.

This of course is a source code (API) break for any existing OCaml
plugins, but we don't guarantee API backwards compatibility for non-C
languages.  Since there are no known external OCaml nbdkit plugins I
only had to update the ones in-tree.