common/include/checked-overflow: provide fallback
On RHEL 7 (GCC 4.8.5) we don't have __builtin_add_overflow and similar
functions.  They were first added in GCC 5.  Provide a fallback path for
these older compilers.

The minimum GCC version we want to support (for the sake of FreeBSD) is
4.2. The fallback path uses the "typeof" and "statement expression" GCC
extensions; those are available in gcc-4.2:

- https://gcc.gnu.org/onlinedocs/gcc-4.2.4/gcc/Statement-Exprs.html
- https://gcc.gnu.org/onlinedocs/gcc-4.2.4/gcc/Typeof.html

The math and the macros were discussed in the following thread:

- [Libguestfs] [PATCH nbdkit 2/2] common/include/checked-overflow.h:
  Provide fallback

  Message-Id: <20211115121417.1174272-2-rjones@redhat.com>

  https://listman.redhat.com/archives/libguestfs/2021-November/msg00177.html

Part of the commit message, and the "configure.ac" hunk, were stolen from
Rich's original patch, linked above.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20211126140652.12662-4-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>